requirements.txt
websockify
socrates_api
django-pipeline
django-pipeline-browserify
